# contracts/openapi/v1/testforge.yaml
openapi: 3.1.0
...
info:
  title: TestForge Control Plane API
  version: 1.0.0
  description: |
    TestForge 项目、用例、Workflow、Run、Worker、Attempt、报告与诊断的公开契约。
    MySQL 是运行状态真相源；派发和回调均按至少一次语义处理。
    AI 诊断仅给出带证据的建议，不改变确定性测试结果。
servers:
  - url: http://localhost:8081
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{projectId}/cases:
    get:
      tags: [Catalog]
      operationId: listCases
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 项目下的原子用例列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TestCase"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Catalog]
      operationId: createCase
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          $ref: "#/components/responses/ResourceCreated"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/cases/{caseId}:
    get:
      tags: [Catalog]
      operationId: getCase
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: 原子用例详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TestCase"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Catalog]
      operationId: updateCase
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "200":
          description: 原子用例已更新
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TestCase"
...
  /api/v1/projects/{projectId}/case-definitions/validate:
    post:
      tags: [Catalog]
      operationId: validateCaseDefinition
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseDefinitionRequest"
      responses:
        "200":
          description: YAML Case 校验结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CaseDefinition"
  /api/v1/projects/{projectId}/case-definitions:
    post:
      tags: [Catalog]
      operationId: createCaseDefinition
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseDefinitionRequest"
      responses:
        "201":
          $ref: "#/components/responses/ResourceCreated"
  /api/v1/cases/{caseId}/definition:
    get:
      tags: [Catalog]
      operationId: getCaseDefinition
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: YAML Case 定义
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CaseDefinition"
    put:
      tags: [Catalog]
      operationId: updateCaseDefinition
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseDefinitionRequest"
      responses:
        "200":
          description: YAML Case 已更新
...
  /api/v1/cases/{caseId}/scripts:
    get:
      tags: [Catalog]
      operationId: listScriptVersions
      parameters:
        - $ref: "#/components/parameters/CaseId"
      responses:
        "200":
          description: 用例的不可变脚本版本列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ScriptVersion"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Catalog]
      operationId: createScriptVersion
      deprecated: true
      description: 仅供旧版初始化与兼容工具；新建和编辑 Case 必须使用 YAML Case Definition API。
      parameters:
        - $ref: "#/components/parameters/CaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScriptVersionRequest"
      responses:
        "201":
          $ref: "#/components/responses/ResourceCreated"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/targets/{targetId}/deployment-profiles:
    get:
      tags: [CICD]
      operationId: listDeploymentProfiles
      parameters:
        - $ref: "#/components/parameters/TargetId"
      responses:
        "200":
          description: 目标的 CI/CD 接入配置
    post:
      tags: [CICD]
      operationId: createDeploymentProfile
      parameters:
        - $ref: "#/components/parameters/TargetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentProfileRequest"
      responses:
        "201":
          description: Jenkins DeploymentProfile 已创建
  /api/v1/targets/{targetId}/deployments:
    get:
      tags: [CICD]
      operationId: listTargetDeployments
      parameters:
        - $ref: "#/components/parameters/TargetId"
      responses:
        "200":
          description: 目标的部署历史
    post:
      tags: [CICD]
      operationId: prepareTargetDeployment
      parameters:
        - $ref: "#/components/parameters/TargetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareDeploymentRequest"
      responses:
        "201":
          description: 已复用或新建不可变 Commit 对应的 Deployment
        "423":
          description: DeploymentProfile 并发配额已满
  /api/v1/deployments/{deploymentId}:
    get:
      tags: [CICD]
      operationId: getTargetDeployment
      parameters:
        - name: deploymentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Deployment 当前状态
  /api/v1/deployments/{deploymentId}/callbacks/jenkins:
    post:
      tags: [CICD]
      operationId: acceptJenkinsDeploymentCallback
      security: []
      parameters:
        - name: deploymentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JenkinsDeploymentCallbackRequest"
      responses:
        "200":
          description: 回调已幂等接受
        "409":
          $ref: "#/components/responses/StateConflict"
...
  /api/v1/runs/{runId}:
    get:
      tags: [Runs]
      operationId: getRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run 聚合及计数
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunSummary"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/reports/runs/{runId}:
    get:
      tags: [Reports]
      operationId: getRunReport
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: 确定性运行报告
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RunReport"
...
  /api/v1/comparison-runs:
    post:
      tags: [Runs]
      operationId: createComparisonRun
      summary: 一次幂等创建 baseline/candidate 配对 Run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComparisonRunRequest"
      responses:
        "201":
          description: 配对 Run 已创建，两侧版本均已固化为 Commit
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ComparisonRun" }
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
  /api/v1/comparison-runs/{comparisonGroupId}:
    get:
      tags: [Runs]
      operationId: getComparisonRun
      parameters:
        - name: comparisonGroupId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      responses:
        "200":
          description: 配对 Run 与两侧固化版本
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ComparisonRun" }
    post:
      tags: [Reports]
      operationId: diagnoseReport
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisRequest"
      responses:
        "200":
          description: 带证据引用的 AI 建议
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/RestResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AiDiagnosis"
        "503":
          description: Provider 不可用；确定性报告仍有效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    TargetId:
      name: targetId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
...
    ResourceCreated:
      description: 资源已创建
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/RestResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ResourceIdentity"
...
    ValidationError:
      description: 字段或状态输入非法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    IdempotencyConflict:
      description: 相同幂等键对应不同内容
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    StateConflict:
      description: 状态或版本已变化
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Uuid:
      type: string
      format: uuid
    RestResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
          pattern: "^[a-f0-9]{32}$"
        data: {}
...
    ResourceIdentity:
      type: object
      required: [id]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        version:
          type: integer
          minimum: 1
...
    CreateCaseRequest:
      type: object
      additionalProperties: false
      required: [name, kind, parameters, timeoutSeconds]
      properties:
        targetId:
          $ref: "#/components/schemas/Uuid"
          description: 兼容字段；项目 Case 省略时由服务端解析当前项目主 Target。
        name:
          type: string
          minLength: 1
          maxLength: 200
        kind:
          type: string
          enum: [ASSERTION, FIXTURE]
        scope:
          type: string
          enum: [PROJECT, SHARED]
          default: PROJECT
        parameters:
          type: object
          additionalProperties: true
        tags:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 64
        timeoutSeconds:
          type: integer
          minimum: 1
          maximum: 86400
        executionRequirement:
          $ref: "#/components/schemas/ExecutionRequirement"
          description: Case 级资源声明；旧客户端省略时按脚本 runner 推导兼容默认值。
...
    CreateScriptVersionRequest:
      type: object
      additionalProperties: false
      required: [runner, sourceRef, checksum]
      properties:
        runner:
          type: string
          enum: [pytest-http, playwright-web, airtest]
        sourceRef:
          type: string
          minLength: 1
          maxLength: 1024
        checksum:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
...
    TestCase:
      type: object
      required: [id, projectId, targetId, name, kind, scope, tags, parameters, timeoutSeconds, createdAt, updatedAt, executionRequirement]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        projectId:
          $ref: "#/components/schemas/Uuid"
        targetId:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        kind:
          type: string
          enum: [ASSERTION, FIXTURE]
        scope:
          type: string
          enum: [PROJECT, SHARED]
        tags:
          type: array
          items:
            type: string
        parameters:
          type: object
          additionalProperties: true
        timeoutSeconds:
          type: integer
        scriptVersionId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        executionRequirement:
          $ref: "#/components/schemas/ExecutionRequirement"
        yamlManaged:
          type: boolean
        definitionDigest:
          type: [string, "null"]
          pattern: "^sha256:[a-f0-9]{64}$"
    CaseDefinitionRequest:
      type: object
      required: [yaml]
      properties:
        yaml:
          type: string
          minLength: 1
    CaseDefinition:
      type: object
      required: [projectId, valid, yaml, digest, name, kind, tags, parameters, timeoutSeconds, executionRequirement, sourceRef, entrypoint, checksum]
      properties:
        caseId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        projectId:
          $ref: "#/components/schemas/Uuid"
        valid:
          type: boolean
        yaml:
          type: string
        digest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        name:
          type: string
        kind:
          type: string
          enum: [ASSERTION, FIXTURE]
        tags:
          type: array
          items:
            type: string
        parameters:
          type: object
          additionalProperties: true
        timeoutSeconds:
          type: integer
        executionRequirement:
          $ref: "#/components/schemas/ExecutionRequirement"
        assetId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        sourceRef:
          type: string
        entrypoint:
          type: string
        checksum:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
...
    ScriptVersion:
      type: object
      required: [id, caseId, runner, sourceRef, checksum, version, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        caseId:
          $ref: "#/components/schemas/Uuid"
        runner:
          type: string
          enum: [pytest-http, playwright-web, airtest]
        sourceRef:
          type: string
        checksum:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
...
    CreateComparisonRunRequest:
      type: object
      additionalProperties: false
      required: [projectId, targetId, environmentId, workflowId, workflowVersion, priority, maxConcurrency, requestKey, baselineRevision, candidateRevision]
      properties:
        projectId: { $ref: "#/components/schemas/Uuid" }
        targetId: { $ref: "#/components/schemas/Uuid" }
        environmentId: { $ref: "#/components/schemas/Uuid" }
        workflowId: { $ref: "#/components/schemas/Uuid" }
        workflowVersion: { type: integer, minimum: 1 }
        priority: { type: integer, minimum: 0, maximum: 9 }
        maxConcurrency: { type: integer, minimum: 1, maximum: 20 }
        processConcurrency: { type: integer, minimum: 1, maximum: 100 }
        deviceConcurrency: { type: integer, minimum: 1, maximum: 100 }
        requestKey: { $ref: "#/components/schemas/Uuid" }
        baselineRevision: { $ref: "#/components/schemas/RevisionSelector" }
        candidateRevision: { $ref: "#/components/schemas/RevisionSelector" }
        deploymentProfileId: { $ref: "#/components/schemas/Uuid" }
    ComparisonRun:
      type: object
      required: [id, requestKey, baselineRun, candidateRun, baselineResolvedCommit, candidateResolvedCommit, createdAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        requestKey: { $ref: "#/components/schemas/Uuid" }
        baselineRun: { $ref: "#/components/schemas/RunSummary" }
        candidateRun: { $ref: "#/components/schemas/RunSummary" }
        baselineResolvedCommit: { type: string, pattern: "^[0-9a-fA-F]{40}$" }
        candidateResolvedCommit: { type: string, pattern: "^[0-9a-fA-F]{40}$" }
        createdAt: { type: string, format: date-time }
    CreateDeploymentProfileRequest:
      type: object
      additionalProperties: false
      required: [name, serverUrl, jobName, buildConfigDigest, maxConcurrency, ttlSeconds]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        serverUrl: { type: string, format: uri, maxLength: 2048 }
        jobName: { type: string, minLength: 1, maxLength: 512 }
        credentialRef:
          type: [string, "null"]
          maxLength: 128
          description: 只保存本机或密钥系统中的引用，不保存 Jenkins Token。
        buildConfigDigest:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        maxConcurrency: { type: integer, minimum: 1, maximum: 50 }
        ttlSeconds: { type: integer, minimum: 60, maximum: 604800 }
    PrepareDeploymentRequest:
      type: object
      additionalProperties: false
      required: [profileId, resolvedCommit]
      properties:
        profileId: { $ref: "#/components/schemas/Uuid" }
        resolvedCommit:
          type: string
          pattern: "^[0-9a-fA-F]{40}$"
    DeploymentState:
      type: string
      enum: [PENDING, TRIGGERED, BUILDING, READY, FAILED, EXPIRED]
    JenkinsDeploymentCallbackRequest:
      type: object
      additionalProperties: false
      required: [callbackKey, providerRunId, status]
      properties:
        callbackKey: { $ref: "#/components/schemas/Uuid" }
        providerRunId: { type: string, minLength: 1, maxLength: 2048 }
        status: { $ref: "#/components/schemas/DeploymentState" }
        endpoint: { type: [string, "null"], maxLength: 2048 }
        artifactUri: { type: [string, "null"], maxLength: 2048 }
        artifactDigest: { type: [string, "null"], maxLength: 255 }
        summary: { type: [string, "null"], maxLength: 1000 }
...
    RunState:
      type: string
      enum: [QUEUED, RUNNING, CANCELLING, SUCCEEDED, FAILED, COMPLETED_WITH_WARNINGS, CANCELLED]
...
    RunSummary:
      type: object
      required: [id, state, totalTasks, taskCounts, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        comparisonGroupId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        testJobId:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        jobConfigVersion:
          type: integer
          minimum: 0
        state:
          $ref: "#/components/schemas/RunState"
        totalTasks:
          type: integer
          minimum: 0
        taskCounts:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: [string, "null"]
          format: date-time
...
    RunReport:
      type: object
      required: [runId, runState, total, passed, failed, pending, durationMs, p50DurationMs, p95DurationMs, failureCategories, results, artifactKeys]
      properties:
        runId:
          $ref: "#/components/schemas/Uuid"
        runState:
          $ref: "#/components/schemas/RunState"
        total: { type: integer, minimum: 0 }
        passed: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
        pending: { type: integer, minimum: 0 }
        durationMs: { type: integer, format: int64, minimum: 0 }
        p50DurationMs:
          type: integer
          minimum: 0
        p95DurationMs:
          type: integer
          minimum: 0
        failureCategories:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
        results:
          type: array
          items:
            $ref: "#/components/schemas/RunReportResult"
        artifactKeys:
          type: array
          items: { type: string }
...
    DiagnosisRequest:
      type: object
      additionalProperties: false
      required: [requestKey, forceRefresh]
      properties:
        requestKey:
          $ref: "#/components/schemas/Uuid"
        forceRefresh:
          type: boolean
...